now it works

+++ engine/warmup.php
@@ -1,19 +1,22 @@
 <?php
+function warmup($date){
 $dir = scandir("../backups");
-$date1 = str_replace(".zip","",$dir[2]);
-$date2 = str_replace(".zip","",$dir[3]);
-if (!empty($date1) && !empty($date2)) {
-if (($date2 - $date1) > 86399) {
- unlink("../backups/".$date1.".zip");
- include_once("backup.php");
- zipData($_SERVER['DOCUMENT_ROOT'], '../backups/'.time().'.zip');
+if(!empty($dir[2])){
+  $fileN = str_replace(".zip","",$dir[2]);
+  if(($fileN -$date) > 86399 )
+  {
+  unlink("../backups/".$dir[2]);
+  include_once("backup.php");
+  zipData($_SERVER['DOCUMENT_ROOT'], '../backups/'.time().'.zip');
 }else{
-unlink("../backups/".$dir[2]);
+  echo "Complete";
 }
-}else{
+}else {
   include_once("backup.php");
   zipData($_SERVER['DOCUMENT_ROOT'], '../backups/'.time().'.zip');
 }
 
+
+}
 
  ?>
